Extract two validation functions to the validators module
